requirements.txt
# Resume parsing
pdfminer.six==20231228
python-docx==1.1.0

# Web scraping & automation
selenium==4.25.0
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
requests==2.32.3

# Scheduling
apscheduler==3.10.4

# NLP for skill extraction
rapidfuzz==3.9.7

# Email
secure-smtplib==0.1.1

# Config
python-dotenv==1.0.1
pyyaml==6.0.2

# Logging
colorlog==6.8.2
